# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BookCreate": {
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Authors",
            "type": "array"
          },
          "cover_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "external_source_id": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Source Id"
          },
          "isbn_10": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 10"
          },
          "isbn_13": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 13"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "metadata_source": {
            "$ref": "#/components/schemas/MetadataSource",
            "default": "manual"
          },
          "page_count": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "published_date": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Date"
          },
          "publisher": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "subtitle": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "BookCreate",
        "type": "object"
      },
...
      "BookResponse": {
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Authors",
            "type": "array"
          },
          "cover_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "external_source_id": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Source Id"
          },
          "household_id": {
            "format": "uuid",
            "title": "Household Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isbn_10": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 10"
          },
          "isbn_13": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 13"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "metadata_source": {
            "$ref": "#/components/schemas/MetadataSource",
            "default": "manual"
          },
          "page_count": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "published_date": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Date"
          },
          "publisher": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "reader_books": {
            "items": {
              "$ref": "#/components/schemas/ReaderBookResponse"
            },
            "title": "Reader Books",
            "type": "array"
          },
          "subtitle": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "title",
          "id",
          "household_id",
          "created_at",
          "updated_at"
        ],
        "title": "BookResponse",
        "type": "object"
      },
...
      "BookUpdate": {
        "properties": {
          "authors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authors"
          },
          "cover_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "isbn_10": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 10"
          },
          "isbn_13": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 13"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "page_count": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "published_date": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Date"
          },
          "publisher": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "subtitle": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "BookUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetadataSource": {
        "enum": [
          "manual",
          "google_books",
          "open_library"
        ],
        "title": "MetadataSource",
        "type": "string"
      },
      "ReaderBookCreate": {
        "properties": {
          "book_id": {
            "format": "uuid",
            "title": "Book Id",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ReadingStatus",
            "default": "planned"
          }
        },
        "required": [
          "book_id"
        ],
        "title": "ReaderBookCreate",
        "type": "object"
      },
      "ReaderBookResponse": {
        "properties": {
          "book_id": {
            "format": "uuid",
            "title": "Book Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ReadingStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reader_id",
          "book_id",
          "status",
          "started_at",
          "finished_at",
          "created_at",
          "updated_at"
        ],
        "title": "ReaderBookResponse",
        "type": "object"
      },
      "ReaderBookUpdate": {
        "properties": {
          "finished_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadingStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ReaderBookUpdate",
        "type": "object"
      },
...
      "ReadingStatus": {
        "enum": [
          "planned",
          "reading",
          "finished"
        ],
        "title": "ReadingStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "My Reading Tracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/books": {
      "get": {
        "operationId": "list_books_api_v1_books_get",
        "parameters": [
          {
            "in": "query",
            "name": "reader_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reader Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReadingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookResponse"
                  },
                  "title": "Response List Books Api V1 Books Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Books",
        "tags": [
          "library"
        ]
      },
      "post": {
        "operationId": "create_book_api_v1_books_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Book",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/books/{book_id}": {
      "delete": {
        "operationId": "delete_book_api_v1_books__book_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "confirm_history",
            "required": false,
            "schema": {
              "default": false,
              "title": "Confirm History",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Book",
        "tags": [
          "library"
        ]
      },
      "get": {
        "operationId": "get_book_api_v1_books__book_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Book",
        "tags": [
          "library"
        ]
      },
      "patch": {
        "operationId": "update_book_api_v1_books__book_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Book",
        "tags": [
          "library"
        ]
      }
    },
...
    "/api/v1/readers/{reader_id}/books": {
      "post": {
        "operationId": "add_book_to_reader_api_v1_readers__reader_id__books_post",
        "parameters": [
          {
            "in": "path",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderBookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReaderBookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Book To Reader",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/readers/{reader_id}/books/{book_id}": {
      "delete": {
        "operationId": "remove_reader_book_api_v1_readers__reader_id__books__book_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Reader Book",
        "tags": [
          "library"
        ]
      },
      "patch": {
        "operationId": "update_reader_book_api_v1_readers__reader_id__books__book_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderBookUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReaderBookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Reader Book",
        "tags": [
          "library"
        ]
      }
    },
...
  }
}